JSBeautify
Used a code formatter to format that code. I use a package for sublime
text. When I am writing code I also have a linter running in the editor
to keep me honest. This is just presence - your code worked fine before
but when you work on a team and especially on larger projects having
standards everyone sticks to does wonders.