add console.logs
I like to use console.log when working as it outputs whatever you give
it. There are times when your function does not run at all and you
would see that with console.log but not assert. Also lets you see the
entire object you are working with so you can see what needs to be
fixed rather than just telling you if the code works or not.